feat: Make an AJAX request in View render function
Add a feature to the View render function to make an asynchronous AJAX request to the Nutritionix API and render the View only after getting the result.
This solution is only temporary and it has been implemented to test that the code was working properly with asynchronous requests in Backbone.