refactor: Move AJAX call to Model
To better separate concerns and separate data and logic from the the presentation layer, moved the AJAX call from View to Model, to be triggered by the `initialize` function inside the latter.
Added code in View's `initialize` function to listen for changes in the Model. Also made `render` inside View in charge of refresh.