refactor: Populate Module URL from options
Refactored code to have the Model URL populated from Model options as a new Model is instantiated. This way the URL, that was previously a fixed string, could be made up of a fixed part plus a query string inserted as options.
This change lets the code move towards the possibility of retrieving that query from a user input in future upgrades.