refactor: Make a Post request to the API
Refactored the code to make a POST request to the API, as illustrated in the Nutritionix API Documentation v1.1. A Model would then be created after the request and then populated with the request results. The code used in this stage was heavily influenced by the code in the Backbone ToDo application developed by Addy Osmani.