refactor: Push models into collection after fetch
Heavily refactored the code to push the models into a collection after having fetched a response from an API. Refactored the Collection delaration file to override its `parse` function to populate a Model's default attributes from the right fields taken from the API response.
Note that this is an incomplete implementation, as it only logs the collection to the console. Everything is working fine, tough. I decided to commit it to freeze the code at this point (that resolved all my issues with fetching and parsing data from an API).
Next commit will have a clean and working implementation with a properly rendered Collection View.