refactor: Render collection
Added code to render models from collection. Depending on the item foods been retrieved from the API and according to the Nutritionix API documentation, serving sizes of food items can be expressed in weight or in quantity.
If the serving size expressed in weight is null, the latter to be expressed in quantity should be provided to the user.