feat: Convert table on small screens
Used two JavaScript libraries to convert the table to a format that would work better on small screens. Of all the various solutions found, Chosed the JavaScript library 'stacktable.js' to format the table.
Used the JavaScript library 'enquire.js' to respond to the media query that would check the screen size.