feat: Add hover effect on table rows
Added hover effect on table rows with the exception of smaller screens. The library the app uses to handle the table transformation for smaller screens makes a new table from each table row. Because of this, being an item properties spread on multiple lines (like rows of a table), hovering over them would only change the appearance of a single line (property).