refactor: Update calories on Collection's items
Made the calories count update only with the item's calories currently in the Collection of selected items. Previously that count was updated whenever a user clicked an item and that item's calories were added, regardless of the fact that that item was already in the Collection.
This calculation has been implemented inside the `food.js` View.