refactor: Create table header dynamically
Changed code to append the table headers dynamically when the corresponding View is created. Made the clear function remove the header each time the render function is called, to avoid duplicating it in the DOM.
Although there is no visible header in the selected items table, it is used by Bootstrap to format correctly the body rows.
The header of the main table is now visible only when there is a selection by the user.